Added option to make it possible to not fail load module for unexisting primitives
When loading the module it is now possible to set the new  Option struct field `disable_strict_load` to true to disable   strict load.
When strict load is enabled loading a module wail behave as previously i.e., load_module fails when it cannot resolve a primitive (for instance it fails for primitives that were not defined using `def_prim` and `install_prim`).
When strict load is disabled the module will not fail when using un existing primitives but instead WARDuino will attach to it a default primitive function called `Primitive_Not_Supported`.
The purpose of disabling strict load is that tool clients can now via `interruptAroundFunction` register new actions to perform once the primitive is encountered during program execution.