added the mock option to the emulator
When starting the emulator a tool client can configure the VM to connect to another target VM (now localhost and port number provided as argument).
This connection is then used to perform remote calls from the emulator VM to the target VM for functions that have been registered through the interruptAroundFunction.