add action and schedule
an action represents an operation that can be execution on intercepted points of interest.
Such points can be for instance around a function call.
For now two possible actions are possible: RemoteCall and ValueSubstitution.

A schedule defines when an action can be triggered.
Some schedulings are based on a timestamp