fix: put pc just at start of instruction and restore
when executing wasm and when the intepreter encounters an instrumented opcode.
The interpretret will automatically increase the pc with 1 byte before jumping to the instrumentor.
The instrumentator thus decreases that pc so that operations like inspect can print the right pc location.
The pc is also restored after all the before actions are executed.